Better examples? (#118)
* Install the bundle_publish workflow always
* Omitting the app deploy workflow for now - maybe we should tie that to
"does an example Dockerfile exist" since we don't at all need it for
infra bundles.
* Instead of doing a bunch of shuffling files around, maybe just include
the Dockerfile in an "examples" folder and let the user decide if they
need it. If they don't, they just delete it, if they do, they'll be
happy, if they just want to look at it, it's not in the way.